update gfs download tools and processing fcns (issues #20, #17, #7)
